:stopwatch: Improved 'nodes' CPU cache usage
I've wanted to do this for a long time, but I was always afraid of the massive code change. But now that nodes are a `std::vector<>`, it was doable.

What's the deal: when you load anything from main memory, even a single byte, a whole cacheline (Common cache line sizes are 32, 64 and 128 bytes) must be fetched to CPU.
When dealing with nodes, most of the time we only access absolute or relative position (mostly just one of those), sometimes mass/velocity/forces and rarely the other parameters.
I figured having to fetch the entire `node_t` to CPU every time to get just one field had to be a bottleneck.
I used VisualStudio profiler and sure enough, the most highlighted lines were those where almost no computation happens but where data need to be read/written from/to memory.

OVERVIEW OF CHANGES:
- look in SimData.h first, it's well highlighted there. Variables `AbsPosition`, `RelPosition`, `Velocity`, `Forces` were moved out from `struct node_t` directly to `class Actor` as separate `std::vector<>`s.
- Actor.h - added the node attributes as standalone `vector<>`s - I preserved the classic names because they stand out and also to make diff more friendly because typos are bound to happen.
- the rest is pretty much just mechanical code rewrite with the help of VisualStudio regex-based find&replace - see below.

Regexes used (refined):
    ar_nodes\[([\S]+)\]\.AbsPosition     -->     ar_nodes_AbsPosition[$1]
    ar_nodes\[([\S]+)\]\.Velocity      -->    ar_nodes_Velocity[$1]
    ar_nodes\[([\S]+)\]\.Forces    -->    ar_nodes_Forces[$1]