:video_game: UI Presets (Novice/Regular/Expert/Minimallist)
Since the game is a very loose sandbox, it's impossible to create an UI to satisfy everyone. But to make some space for debate and future enhancements, I've coined these 4 presets.

Suggested on Discord: https://discord.com/channels/136544456244461568/189904947649708032/1128168306558390383

I took care to make it data-driven and intuitively readable in code, so that users can read it as reference and UI automatically reflects changes.